SRE_4/fastapi-microservices-master/fastapi-microservices-master/order-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
requests==2.32.3
prometheus-fastapi-instrumentator
